Create hoip_array.sh
This bash script takes a folder of .car and .mdf (with a lammps input file, slurm job script, and msi2lmp.exe) files and gets lammps simulations running. 
To do this, it creates a list of filenames from the .car .mdf file extensions, then converts those to a lammps .data file using msi2lmp.exe. It then creates a folder using the basename of the .car .mdf pair, and moves the .car, .mdf, and .data file into it. 
Afterwards, it uses the lammps input and slurm script to start all the jobs.